refactor(front): move LogsData to prisma folder
