+++ front/src/app/page.tsx
@@ -1,6 +1,6 @@
 import { FilterHeader } from "@/features/filter/FilterHeader";
 import { TableBody } from "@/features/tables/views/TableBody";
-import { LogsData } from "@/services/LogsData";
+import { LogsData } from "@/services/prisma/LogsData";
 
 export const revalidate = 0;
